+++ test/TestTypeResolver.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using System.Text.Json.Serialization.Metadata;
+using Tavenem.DataStorage;
+
+namespace Tavenem.DataStore.Test;
+
+public class TestTypeResolver : DefaultJsonTypeInfoResolver
+{
+    /// <inheritdoc />
+    public override JsonTypeInfo GetTypeInfo(Type type, JsonSerializerOptions options)
+    {
+        var jsonTypeInfo = base.GetTypeInfo(type, options);
+
+        if (jsonTypeInfo.Type == typeof(IIdItem))
+        {
+            var derivedType = new JsonDerivedType(typeof(TestIdItem), TestIdItem.TestIdItemTypeName);
+            if (jsonTypeInfo.PolymorphismOptions is null)
+            {
+                jsonTypeInfo.PolymorphismOptions = new JsonPolymorphismOptions
+                {
+                    IgnoreUnrecognizedTypeDiscriminators = true,
+                    UnknownDerivedTypeHandling = JsonUnknownDerivedTypeHandling.FallBackToNearestAncestor,
+                    DerivedTypes = { derivedType },
+                };
+            }
+            else
+            {
+                jsonTypeInfo.PolymorphismOptions.DerivedTypes.Add(derivedType);
+            }
+        }
+
+        return jsonTypeInfo;
+    }
+}